Rework to attempt building layer hierarchies
Build a resolution graph based on OSGi wirings that can detect split
packages and cycles.  For modules that are not involved with split
packages or cycles we attempt to build a JPMS layer hierarchy for the
bundle module.

This has advantages because it allows us to model require bundle 
reexports as JPMS requires transitive properly.

Change-Id: I716c30cd394a701bcf05f2b20663a44e89307ced
Signed-off-by: Thomas Watson <tjwatson@us.ibm.com>